creating a basic FIFO stack using ingeritance
