+++ CSharpFundamentalsSandbox/Katas/OOP/Classes/Post.cs
@@ -1,0 +1,48 @@
+using System;
+
+namespace CSharpFundamentalsSandbox.Katas.OOP.Classes
+{
+    public class Post
+    {
+        // Exercise 2
+        // Design a class called Post. This class models a StackOverflow post.
+        // It should have properties for title, description and the date/time it was created.
+        // We should be able to up-vote or down-vote a post.
+        // We should also be able to see the current vote value.
+        // In the main method, create a post, up-vote and down-vote it a few times and then display the current vote value.
+        // In this exercise, you will learn that a StackOverflow post should provide methods for up-voting and down-voting.
+        // You should not give the ability to set the Vote property from the outside, because otherwise, you may accidentally change the votes of a class to 0 or to a random number.
+        // And this is how we create bugs in our programs. The class should always protect its state and hide its implementation detail
+
+
+        public Post(string title, string description)
+        {
+            Title = title;
+            Description = description;
+            CreatedAt = DateTime.Now;
+        }
+
+        public int NumberOfVotes { get; private set; }
+        public string Title { get; set; }
+        public string Description { get; set; }
+        public DateTime CreatedAt { get; set; }
+
+        public void UpVote()
+        {
+            NumberOfVotes += 1;
+        }
+
+        public void DownVote()
+        {
+            NumberOfVotes -= 1;
+        }
+
+        public void PostDetails()
+        {
+            Console.WriteLine("Title: {0}", Title);
+            Console.WriteLine("Description: {0}", Description);
+            Console.WriteLine("Date Posted: {0}", CreatedAt);
+            Console.WriteLine("Number of votes: {0}", NumberOfVotes);
+        }
+    }
+}
