+++ CSharpFundamentalsSandbox/Katas/OOP/Classes/StopWatch.cs
@@ -1,0 +1,46 @@
+using System;
+
+namespace CSharpFundamentalsSandbox.Katas.OOP.Classes
+{
+    public class StopWatch
+    {
+        // EXERCISE 1
+        // Design a class called Stopwatch. The job of this class is to simulate a stopwatch.
+        // It should provide two methods:
+        // Start and Stop. We call the start method first, and the stop method next.
+        // Then we ask the stopwatch about the duration between start and stop.
+        // Duration should be a value in TimeSpan. Display the duration on the console. We should also be able to use a stopwatch multiple times.
+        // So we may start and stop it and then start and stop it again. Make sure the duration value each time is calculated properly.
+        // We should not be able to start a stopwatch twice in a row (because that may overwrite the initial start time). So the class should throw an InvalidOperationException if its started twice.
+
+        private DateTime _startTime;
+        private bool _timerStarted;
+
+        public void StartTimer()
+        {
+            if (!_timerStarted)
+            {
+                _timerStarted = true;
+                _startTime = DateTime.Now;
+                Console.WriteLine("Timer Started");
+            }
+            else
+            {
+                throw new InvalidOperationException("StartTimer is already running");
+            }
+        }
+
+        public void StopTimer()
+        {
+            if (_timerStarted)
+            {
+                _timerStarted = false;
+                Console.WriteLine("Timer was running for {0}", DateTime.Now - _startTime);
+            }
+            else
+            {
+                throw new InvalidOperationException("StartTimer must be called first");
+            }
+        }
+    }
+}
