+++ CSharpFundamentalsSandbox/Program.cs
@@ -1,32 +1,42 @@
-using static CSharpFundamentalsSandbox.Katas.OOP.Classes;
+using System;
+using CSharpFundamentalsSandbox.Katas.OOP.Inheritance;
 
 namespace CSharpFundamentalsSandbox
 {
     internal class Program
     {
-        // private static void Helper(Action action)
-        // {
-        //     try
-        //     {
-        //         action();
-        //     }
-        //     catch (Exception e)
-        //     {
-        //         Console.WriteLine(e);
-        //     }
-        // }
+        private static void Helper(Action action)
+        {
+            try
+            {
+                action();
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
+        }
 
         public static void Main(string[] args)
         {
-            var post = new Post("Test Title", "Test Description");
-            post.UpVote();
-            post.UpVote();
-            post.UpVote();
-            post.UpVote();
-            post.UpVote();
-            post.DownVote();
+            var stack = new Stack();
 
-            post.PostDetails();
+            Helper(() => stack.Push("1"));
+            // Helper((() => stack.Push(2)));
+            // Helper((() => stack.Push(3)));
+            // Helper((() => stack.Push(4)));
+            //
+            // Helper((() => stack.Pop()));
+            // Helper((() => stack.Pop()));
+            // Helper((() => stack.Pop()));
+
+            // Helper((() => stack.Clear()));
+            //
+            // Helper((() => stack.Clear()));
+        }
+
+        public class TestClassGeneration
+        {
         }
     }
 }
